+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
-## [0.11.0]  2022-04-10
+## [0.11.0]  2022-04-11
 ### Changes
-- table data builder change to array of FromItemInfo
+- Table data builder change to array of FromItemInfo
+- testTable in unit test renamed to table1
+- Remove set builderOption from OrderByItemInfo and SelectItemInfo, not needed anymore as we are passing the builderData to getStmt()
+- Update RailRoad documentation to include Joins steps
 
 ### Added
 - From step can take more than one table
 - Builder option addAsBeforeTableAlias to Builder Option
 - AliasedTable class, table can be aliased
 - CrossJoin Step
+- Join, InnerJoin, LeftJoin, RightJoin, FullOuterJoin Steps
 
 ## [0.10.1]  2022-04-08
 ### Changes